% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_screening.R
\name{plot.nntmvn_1stmmt_pred}
\alias{plot.nntmvn_1stmmt_pred}
\title{Plot function for the \code{nntmvn_1stmmt_pred} class}
\usage{
\method{plot}{nntmvn_1stmmt_pred}(x)
}
\value{
a ggplot object of class "gg" and "ggplot"
}
\description{
Plot function for the \code{nntmvn_1stmmt_pred} class
}
