% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmvn_sample.R
\name{seq_Vecc_samp_func}
\alias{seq_Vecc_samp_func}
\title{Simulate truncated multivariate normal (TMVN) as censored locations using
nearest neighbors}
\usage{
seq_Vecc_samp_func(
  y,
  cens_lb,
  cens_ub,
  mask_cens,
  NN,
  locs,
  cov_name,
  cov_parm,
  covmat = NULL,
  method = c("exact", "Vecchia")
)
}
\arguments{
\item{y}{uncensored responses of length n, where n is the number of locations}

\item{cens_lb}{lower bound vector for TMVN of length n}

\item{cens_ub}{upper bound vector for TMVN of length n}

\item{mask_cens}{mask for censored responses (also locations) of length n}

\item{NN}{n X m matrix for nearest neighbors. i-th row is the nearest neighbor indices of y_i. \code{NN[i, 1]} should be \code{i}}

\item{locs}{location matrix n X d}

\item{cov_name}{covariance function name from the \code{GpGp} package}

\item{cov_parm}{parameters for \code{cov_name}}

\item{covmat}{(optional) n-by-n dense covariance matrix}
}
\value{
a vector of length n with imputed results for censored responses
}
\description{
Simulate truncated multivariate normal (TMVN) as censored locations using
nearest neighbors
}
